Reverse up-down viewport navigation from Q and E keys (#115)
Reversing the up-down viewport camera translation navigation so that Q makes the camera go down and E makes the camera to go up (earlier they were swapped). This is to match the original Nerfstudio viewer (and also similar to Unity).

Regarding the up-down arrow keys for panning rotating up and down, I noticed that they are reversed compared to the earlier nerfstudio viewer, however the current direction of movement for the up down arrow keys in viser is more ideal for viewing meshes. I am wondering if there is a way to only customize the key input for the nerfstudio viewer without affecting viser for only the up down arrow keys.