Make 'gdown' an example group dependency.
gdown is only used to fetch the assets used for examples -
colmap_garden, dragon_mesh, and record3d_dance (scripts in
`examples/assets/`).

It is a CLI tool and not used anywhere in the main codebase.
This PR updates `pyproject.toml` and moves gdown to the optional
dependency group. It shortens the dependency tree of this library
and hence, benefits downstream projects.